+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Revision history for futhark-server
 
-## 0.1.0.0 -- 2021-06-17
+## 1.1.0.0 -- 2021-07-01
+
+* `cmdInputs` and `cmdOutputs` now return `InputType` and `OutputType`
+  values instead of just `TypeName`, in order to also capture
+  consumption information.
+
+## 1.0.0.0 -- 2021-06-17
 
 * First version. Released on an unsuspecting world.